requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
certifi
colorama
deprecation
et_xmlfile
frozenlist
gotrue
h11
h2
hpack
httpcore
httpx
hyperframe
idna
iniconfig
multidict
numpy
openpyxl
packaging
pandas
pluggy
postgrest
propcache
pydantic
pydantic_core
Pygments
PyJWT
pytest
pytest-mock
python-dateutil
python-dotenv
python-telegram-bot
pytz
realtime
six
sniffio
storage3
StrEnum
supabase
supafunc
# typing-inspection
# typing_extensions
# tzdata
# websockets
# yarl
